+++ check-reserve-update-frequency/src/bot-setup.js
