+++ compare-oracle-to-fallback/src/bot-setup.js
