+++ total-value-and-liquidity/src/bot-setup.js
